requirements.txt
h5py==2.7.0
Keras==2.0.4
matplotlib==2.0.2
numpy==1.12.1
pandas==0.20.2
seaborn==0.7.1
sklearn==0.0
tensorflow-gpu==1.1.0
